third_party/blink/web_tests/fast/events/init-message-event.html

<!DOCTYPE html>
<html>
<head>
<link rel="help" href="http://www.w3.org/TR/2013/WD-html51-20130528/editing.html#dom-datatransferitemlist-remove">
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Checks that initMessageEvent() is working");

var newEvent;
function runTest() {
    newEvent = document.createEvent("MessageEvent");
    shouldThrow('newEvent.initMessageEvent("message", true, true, "ff", "*", 43, window, {x:1});', '"TypeError: Failed to execute \'initMessageEvent\' on \'MessageEvent\': Iterator getter is not callable."');
    shouldNotThrow('newEvent.initMessageEvent("message", true, true, "ff", "*", 43, window, undefined)');
    shouldBeType(newEvent, MessageEvent);
}

runTest();
</script>
</body>
</html>
